+++ app/src/main/java/com/example/deliveryapp/model/Restaurant.kt
@@ -1,3 +1,3 @@
-package com.example.deliveryapp.models
+package com.example.deliveryapp.model
 
 data class Restaurant(val name: String, val location: String)
